\name{formula.bqtl}
\alias{formula.bqtl}

\title{ Extract formula from bqtl object }
\description{
  formula method for class \code{bqtl}
}
\synopsis{formula.bqtl(x,...)}
\usage{
formula.bqtl(x)
}

\arguments{
  \item{x}{The object returned by \code{bqtl} }
}

\value{ a formula object
}

\author{Charles C. Berry \email{cberry@ucsd.edu} }
\seealso{\code{\link{bqtl}}}
\keyword{methods}
