\name{little.ana.f2}
\docType{data}
\alias{little.ana.f2}
\title{ A simulated dataset }
\description{
A simulation of an F2 cross of 150 organisms with a genome of around 500
cM consisting of 5 chromosomes. The format is that created by \code{\link{make.analysis.obj}}
}
\usage{data(little.ana.f2)}
\keyword{datasets}