\name{little.bc.markers}
\alias{little.bc.markers}
\non_function{}
\title{ Simulated Marker Data }
\description{
  The \code{little.bc.markers} data frame has 150 rows and 50
  columns with the simulated marker data from a BC1 cross of 150
  organisms with a genome of around 500 cM consisting of 5
  chromosomes. Some \code{NA}'s have been intentionally introduced. 
}
\format{
This data frame contains the following columns:
\describe{
\item{m.1}{
a factor with levels
\code{AA} 
\code{Aa} 
}
\item{m.2}{
a factor with levels
\code{AA} 
\code{Aa} 
}
\dots
\item{m.49}{
a factor with levels
\code{AA} 
\code{Aa} 
}
\item{m.50}{
a factor with levels
\code{AA} 
\code{Aa} 
}
}
}

\keyword{datasets}
