\name{little.bc.pheno}
\docType{data}
\alias{little.bc.pheno}
\title{ Simulated Phenotype Data  }
\description{
The \code{little.bc.pheno} data frame has 150 rows and 1 columns.

}
\format{
This data frame contains the following columns:
\describe{
\item{bc.phenotype}{
  a numeric vector of simulated phenotype data}}}
\usage{data(little.bc.pheno)}
\keyword{datasets}
