\name{little.map.dx}
\alias{little.map.dx}
\non_function{}
\title{ Marker Map Description for Simulated Data }
\description{
The \code{little.map.dx} data frame has 50 rows and 2 columns that specify 
 the map locations of a simulated set of markers }

\format{
This data frame contains the following columns:
\describe{
\item{marker.names}{
a factor with levels
\code{m.1} \dots \code{m.50} 
}
\item{cM}{
a numeric vector of map locations in centimorgans
}
}
}  
\keyword{datasets}