\name{little.f2.markers}
\alias{little.f2.markers}
\non_function{}
\title{ Simulated Marker Data }
\description{
The \code{little.f2.markers} data frame has 150 rows and 50 columns with
the simulated marker data from an F2 cross of 150
  organisms with a genome of around 500 cM consisting of 5 chromosomes.
}
\format{
This data frame contains the following columns:
\describe{
\item{m.1}{
a factor with levels
\code{AA} 
\code{Aa}
\code{aa} 
}
\item{m.2}{
a factor with levels
\code{AA} 
\code{Aa}
\code{aa}  
}
\dots
\item{m.25}{
a factor with levels
\code{A-} 
\code{aa} 
}
\dots
\item{m.45}{
a factor with levels
\code{a-} 
}
\dots
\item{m.49}{
a factor with levels
\code{AA} 
\code{Aa} 
\code{aa}
}
\item{m.50}{
a factor with levels
\code{AA} 
\code{Aa} 
\code{aa}
}
}
}

\keyword{datasets}
