\name{little.f2.pheno}
\docType{data}
\alias{little.f2.pheno}
\title{ Simulated Phenotype Data }
\description{
The \code{little.f2.pheno} data frame has 150 rows and 1 columns.

}
\format{
This data frame contains the following columns:
\describe{
\item{f2.phenotype}{
a numeric vector of simulated phenotype data
}
}
}
\usage{data(little.f2.pheno)}
\keyword{datasets}
