\name{loglik}
\alias{loglik}
\alias{loglik.default}
\alias{loglik.bqtl}
\alias{loglik.bqtl.list}
\alias{logpost}
\alias{logpost.default}
\alias{logpost.bqtl}
\alias{logpost.bqtl.list}
\alias{posterior}
\alias{posterior.default}
\alias{posterior.bqtl}
\alias{posterior.bqtl.list}
\title{ Extract loglikelihood, log posterior, or posterior from fitted models }
\description{
A fitted model or a list of such generated by \code{bqtl} has a maximum
log likelihood or log posterior and a posterior. These functions simply
extract them.
}
\usage{
loglik(x, ...)
logpost(x, ...)
posterior(x, ...)
}

\arguments{
  \item{x}{ The object produced by \code{bqtl} }
  \item{...}{ Currently unused }
}

\value{
 A vector of numbers whose length equals the number of fitted models in \code{x}
}
\author{Charles C. Berry \email{cberry@ucsd.edu} }
\seealso{ \code{\link{bqtl}} }
\keyword{models}