\name{summary.map.frame}
\alias{summary.map.frame}
\alias{summary.analysis.object}

\title{ Summary methods for basic data objects }
\description{
 Provide a simple report on the data structure
}
\usage{
\method{summary}{map.frame}(object,\dots)
}

\arguments{
  \item{object}{ A \code{map.frame} or \code{analysis.object} }
  \item{\dots}{ignored}
}

\value{
 a list
}
\author{Charles C. Berry \email{cberry@ucsd.edu} }
\keyword{methods}
