\name{bqtl-internal}
\alias{\%equiv\%}
\alias{map.dx}
\alias{rhs.bqtl}
\alias{unique.config}
\alias{zero.dup}
\alias{version.bqtl}
\title{Internal BQTL functions}
\description{
 Internal bqtl functions and objects
}
\usage{
x\%equiv\%y
map.dx(lambda)
rhs.bqtl(reg.terms, ana.obj, bqtl.specials, local.covar, scope, 
    expand.specials = NULL, method, ...) 
zero.dup(x)
unique.config(swap.obj)

}
\details{
  These are not to be called by the user.
}
\keyword{utilities}
