\name{formula.bqtl}
\alias{formula.bqtl}

\title{ Extract formula from bqtl object }
\description{
  formula method for class \code{bqtl}
}
\usage{
\method{formula}{bqtl}(x, \dots)
}

\arguments{
  \item{x}{The object returned by \code{bqtl} }
  \item{\dots}{ unused }
}

\value{ a formula object
}

\author{Charles C. Berry \email{cberry@ucsd.edu} }
\seealso{\code{\link{bqtl}}}
\keyword{methods}
