\name{little.f2.markers}
\alias{little.f2.markers}
\non_function{}
\title{ Simulated Marker Data }
\description{
The \code{little.f2.markers} data frame has 150 rows and 50 columns with
the simulated marker data from an F2 cross of 150
  organisms with a genome of around 500 cM consisting of 5 chromosomes.
}
\format{
This data frame contains the following columns:
\describe{
\item{m.1}{
a factor with levels
\code{AA} 
\code{Aa}
\code{aa} 
}
\item{m.2}{
a factor with levels
\code{AA} 
\code{Aa}
\code{aa}  
}
\item{m.3}{ ditto }
 \item{m.4}{ ditto }
 \item{m.5}{ ditto }
 \item{m.6}{ ditto }
 \item{m.7}{ ditto }
 \item{m.8}{ ditto }
 \item{m.9}{ ditto }
 \item{m.10}{ ditto }
 \item{m.11}{ ditto }
 \item{m.12}{ ditto }
 \item{m.13}{ ditto }
 \item{m.14}{ ditto }
 \item{m.15}{ ditto }
 \item{m.16}{ ditto }
 \item{m.17}{ ditto }
 \item{m.18}{ ditto }
 \item{m.19}{ ditto }
 \item{m.20}{ ditto }
 \item{m.21}{ ditto }
 \item{m.22}{ ditto }
 \item{m.23}{ ditto }
 \item{m.24}{ ditto }
\item{m.25}{
a factor with levels
\code{A-} 
\code{aa} 
}
 \item{m.26}{ ditto }
 \item{m.27}{ ditto }
 \item{m.28}{ ditto }
 \item{m.29}{ ditto }
 \item{m.30}{ ditto }
 \item{m.31}{ ditto }
 \item{m.32}{ ditto }
 \item{m.33}{ ditto }
 \item{m.34}{ ditto }
 \item{m.35}{ ditto }
 \item{m.36}{ ditto }
 \item{m.37}{ ditto }
 \item{m.38}{ ditto }
 \item{m.39}{ ditto }
 \item{m.40}{ ditto }
 \item{m.41}{ ditto }
 \item{m.42}{ ditto }
 \item{m.43}{ ditto }
 \item{m.44}{ ditto }
\item{m.45}{
a factor with levels
\code{a-} 
}
\item{m.46}{ ditto }
 \item{m.47}{ ditto }
 \item{m.48}{ ditto }
\item{m.49}{
a factor with levels
\code{AA} 
\code{Aa} 
\code{aa}
}
\item{m.50}{
a factor with levels
\code{AA} 
\code{Aa} 
\code{aa}
}
\item{row.names}{row names}}
}

\keyword{datasets}
