\name{little.f2.pheno}
\alias{little.f2.pheno}
\non_function{}
\title{ Simulated Phenotype Data }
\description{
The \code{little.f2.pheno} data frame has 150 rows and 1 columns.

}
\format{
This data frame contains the following columns:
\describe{
\item{f2.phenotype}{
a numeric vector of simulated phenotype data
}
}
}

\keyword{datasets}
