\name{summary.map.frame}
\alias{summary.map.frame}
\alias{summary.analysis.object}

\title{ Summary methods for basic data objects }
\description{
 Provide a simple report on the data structure
}
\usage{
summary(x)

summary.map.frame(x)

summary.analysis.object(x)
}

\arguments{
  \item{x}{ A \code{map.frame} pr \code{analysis.object} }
}

\value{
 a list
}
\author{Charles C. Berry \email{cberry@ucsd.edu} }
\keyword{bqtl}
